1	ocm and ocmm don't work for one covariate.  2	ocm and ocmm don't work if the data argument isn't specified (even if all variables in  model are in the workspace). Thanks to find out these two! It works now as expected.
 3	(Not a bug) In plot.ocmm help, we have R=number of bootstrap replicates, but bootstrap CIs aren't provided ??
Ok, maybe this requires a few words. R is used in the method ‘vcov'. The vcov matrix is used to randomly generate R sets of parameters (sample from a multivariate normal) and then R actual g functions. The CIs are then computed as 95% quantiles of the R g functions for each value of v. We could explain all this, or decide to use a fixed number for the ‘vcov’ function, say Rvcov=1000, not giving the user the possibility to choose.

 4	extractAIC returns a vector with 2 elements (df,AIC). This is not what the helpfile says.
Fixed in the documentation.

 5	logLik returns a list with 5 elements.This is not what the helpfile says.
Technically it is a number with a few attributes (it is a requirement for any logLik method). I have fixed the documentation.

 6	(Not a bug) The bootstrap CIs take a long time.
 7	(Not a bug) ocmm is quite slow (I used n=1000).
Yes it is a bit slow. My idea was to increase performances in the next releases.
Anyway, I have added the @param “method” to ocm and ocmm . The default is “optim” (what we were using), but now the user can also choose “ucminf” (from package “ucminf”- General-purpose unconstrained non-linear optimization, used in “ordinal”). This gives quite a speed boost (2x).
In the file attached I have added a few speed tests at the end. Ordinal is still much faster, but the gap is decreasing and anyway we are using continuous scales, while ordinal uses discrete scales. There is space for improvement, but I would leave this to future versions.